# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Navidrome Replacement API
  version: 0.1.0
  description: Self-hosted music server API
servers:
  - url: http://localhost:8090
    description: Local development
...
paths:
...
  /api/v1/tracks/{trackId}/stream:
    get:
      tags: [playback]
      operationId: streamTrack
      summary: Stream track audio with HTTP Range support
      parameters:
        - $ref: "#/components/parameters/trackId"
      responses:
        "200":
          description: Full audio stream
          content:
            audio/*:
              schema:
                type: string
                format: binary
        "206":
          description: Partial content (Range request)
          content:
            audio/*:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "416":
          $ref: "#/components/responses/RangeNotSatisfiable"
  /api/v1/playlists:
    get:
      tags: [playlists]
      operationId: listPlaylists
      summary: List playlists
      responses:
        "200":
          description: Playlist list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistList"
    post:
      tags: [playlists]
      operationId: createPlaylist
      summary: Create playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistCreate"
      responses:
        "201":
          description: Created playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/playlists/{playlistId}:
    get:
      tags: [playlists]
      operationId: getPlaylist
      summary: Get playlist with tracks
      parameters:
        - $ref: "#/components/parameters/playlistId"
      responses:
        "200":
          description: Playlist detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RangeNotSatisfiable:
      description: Range not satisfiable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
...
    trackId:
      name: trackId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    playlistId:
      name: playlistId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code, message]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
...
    Track:
      type: object
      required: [id, title, artistName, albumId, durationMs, format]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artistName:
          type: string
        albumId:
          type: string
          format: uuid
        albumTitle:
          type: string
        trackNo:
          type: integer
        durationMs:
          type: integer
        format:
          type: string
        genre:
          type: string
        sampleRateHz:
          type: integer
        bitDepth:
          type: integer
        bitrateKbps:
          type: integer
        sizeBytes:
          type: integer
...
    Playlist:
      type: object
      required: [id, name, isDefault, trackCount]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isDefault:
          type: boolean
        trackCount:
          type: integer
    PlaylistList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Playlist"
        total:
          type: integer
    PlaylistDetail:
      allOf:
        - $ref: "#/components/schemas/Playlist"
        - type: object
          required: [tracks]
          properties:
            tracks:
              type: array
              items:
                $ref: "#/components/schemas/Track"
    PlaylistCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
...